API

Documentation for Wired.jl's public-facing API.

Sources

Define custom source types.

Wired.Source — Type

abstract type Source

Represents a time-invariant current source that generates a magnetic field

Wired.Wire — Type

struct Wire <: Source

Represents a finite wire segment in 3D space

The wire segment has a finite radius, such that numerical singularities are prevented: the magnetic field generated by the wire segment has a finite (maximum) value on the surface at r=R, and decays linearly to zero at the center with r=0.

Fields

a0::Vector{Float64}: XYZ coordinates of the start of the wire vector

a1::Vector{Float64}: XYZ coordinates of the end of the wire vector

I::Float64: total current in the filament

R::Float64: radius of the wire segment

Wired.Ring — Type

abstract type Ring <: Source

A circular current-carrying ring with circular or rectangular-cross-section.

Major axis of the ring is the Z-axis.

Wired.CircularRing — Type

struct CircularRing <: Ring

A circular current-carrying solid conducting ring with circular cross-section

The ring has a defined cross-section, such that numerical singularities are prevented and the magnetic field can be accurately calculated within the ring itself.

Fields

name::String: describes the filament

H::Float64: distance along the Z-axis of the ring from the coordinate system centroid to the ring centroid

R::Float64: major radius of the ring

r::Float64: minor radius of the ring (cross-section radius)

I::Float64: current in the conducting ring; sign convention follows right-hand rule about primary axis.

Wired.RectangularRing — Type

struct RectangularRing <: Ring

A circular current-carrying solid conducting ring with rectangular cross-section

The ring has a defined cross-section, such that numerical singularities are prevented and the magnetic field can be accurately calculated within the ring itself.

Fields

name::String: describes the filament

H::Float64: distance along the Z-axis of the ring from the coordinate system centroid to the ring centroid

R::Float64: major radius of the ring

w::Float64: width of the ring cross-section

h::Float64: height of the ring-cross-section

I::Float64: current in the conducting ring; sign convention follows right-hand rule about primary axis.

Mathematical Functions

Define high-performance version of elliptic integral and linear algebra functions optimized for use with this code.

Wired.ellipK — Function

function ellipK(k2; itmax=100, errmax=1e-8)

Solve the complete elliptic integral of the first kind

Wired.ellipE — Function

function ellipE(k2; itmax=100, errmax=1e-8)

Solve the complete elliptic integral of the second kind

Reference: https://mathworld.wolfram.com/CompleteEllipticIntegraloftheSecondKind.html

Wired.ellipKE — Function

ellipKE(k2; itmax=100, errmax=1e-12)

Solve the complete elliptic integrals of the first and second kinds.

Wired.crossrows! — Function

crossrows!(C::AbstractArray, A::AbstractArray, b::Vector)

Cross each row of A by a 3-length vector b and place in C, i.e.

C[i,:] = cross(A[i,:], b)

Wired.normrows! — Function

normrows!(b::Vector, A::AbstractArray)

Calculate the vector norm of each row in Nx3 matrix and place in b, i.e.

b[i] = norm(A[i,:])

Wired.multrows! — Function

multrows!(A::AbstractArray, b::Vector)

In-place element-wise multiplication of each row of an Nx3 matrix by the corresponding element in vector b, i.e.

A[i,:] .*= b[i]

Wired.dotrows! — Function

dotrows!(c::Vector, A::AbstractArray, b::Vector)

In-place dot-product of the rows of A by vector b; place in C, i.e.

c[i] .*= dot(A[i,:], b)       # or C[i,:] .*= sum(A[i,:] .* b)

Processing Routines

Wired.makewires — Function

makewires(mesh::Mesh)

Create finite Wire segment sources from a Mesh

Arguments

mesh::Mesh: the finite element mesh to generate Wires from

ratio=2.0: the length/radius ratio for the wire segments. Set to 2.0 for perfect cubes and 1.0 for tetrahedrons.

Returns

Vector{Wire} containing all Wire objects in the mesh

Wired.makecircrings — Function

function makecircrings(rect::RectangularRings; Nmin=2)

Filamentize Rings with rectangular cross-sections.

For a cross-section with side lengths L1 and L2, with L1 < L2, the spacing s between filament centroids is given by s = L1 / Nmin. This spacing is used for creating filaments along both sides, such that N1 = Nmin and N2 ~= L2 / s (with integer rounding.)

Arguments

rect::Vector{RectangularRing}: RectangularRing objects from which to create CircularRing objects

'Nmin=2`: Minimum number of CircularRings per side of the original cross-section.

Returns

Vector{CircularRing}, which is of indeterminate length apriori.

Wired.threadindices — Function

threadindices(it::Integer, Nt::Integer, N::Integer)

Split a problem up for multi-threaded operation

For thread number it, total number of threads Nt, and total number of tasks N, determine the start/stop index for that particular thread number

File I/O

Define functions for loading and save data to/from files.

Wired.loadmesh — Function

loadmesh(fn::String)

Load a finite element mesh from a file

The first row is a header row. The subsequent rows are: Xcentroid, Ycentroid, Zcentroid, volume, JdensityX, JdensityY, JdensityZ

Jdensity is optional (the file may be 4-width)

Wired.savemesh — Function

savemesh(fn::String, mesh::Mesh)

Save a finite element mesh to a file.

Wired.loadrings — Function

loadrings(fn::String)

Load Ring objects from file.

Returns

Vector{CircularRing}, Vector{RectangularRing}

Wired.saverings — Function

saverings(fn::String, rings::Vector{<:Source})

Save Ring objects to a file

Wired.loadwires — Function

loadwires(fn::String)

Load Wire objects from a file

Wired.savewires — Function

savewires(fn::String, wires::Vector{Wire})

Save Wire objects to a file

Output Fields

Define functions that create basic sets of points at which to calculate fields.

Wired.Line — Type

line(start, stop, N)

Represents points along a line in 3D space

Fields

start::Vector{Float64}: 3-length start point of the line

stop::Vector{Float64}: 3-length end point of the line

N::Integer: number of points along the line

nodes::Matrix{Float64}: Nx3Matrix` of locations for calculations

Wired.Mesh — Type

mutable struct Mesh

Define a finite element mesh, which can act as a Source and a Field

Fields

nodes::AbstractArray: Nx3 matrix of positions in 3D space (element centroids)

volumes::AbstractArray: N-length vector of volume of each element

Jdensity::AbstractArray: Nx3 matrix of current density vectors at each element

mu_r::Float64: Relative magnetic permeability of the mesh

Biot-Savart Solvers

Define functions that calculate the magnetic flux density due to defined current sources.

Wired.bfield — Function

bfield(nodes::AbstractArray, rings::Vector{Ring}; Nmin=2, errmax=1e-8, Nt=0)

Calculate the B-field at a collection of points in 3D space, generated by a series of Ring objects.

Arguments

nodes::AbstractArray: Nx3 Matrix containing (x,y,z) coordinates of points in 3D space

wires::Vector{<:Ring}: Ring objects contributing to the magnetic field (circular or rectangular cross-section)

Nmin::Integer: minimum number of CircularRing objects to use to represent the shortest edge of a rectangular cross-section

errmax::Float64: maximum error tolerance for elliptic integral calculations

Nt::Integer: number of threads to use for the calculation (default: all available threads)

Returns

Nx3 Matrix containing magnetic flux density vectors at each of the points in 3D space represented by nodes

bfield(nodes::AbstractArray, wires::Vector{Wire}; 
        Nt::Integer=0, mu_r=1.0)

Calculate the B-field at a collection of points in 3D space, generated by a series of finite-length Wire objects.

Arguments

nodes::AbstractArray: Nx3 Matrix containing (x,y,z) coordinates of points in 3D space

wires::Vector{Wire}: Wire objects contributing to the magnetic field

Nt::Integer: number of threads to use for the calculation (default: all available threads)

Returns

Nx3 Matrix containing magnetic flux density vectors at each of the points in 3D space represented by nodes